Fix data race in fitness file write for robot
    In Robot.Get_Fitness(), the fitness is first written to a
temporary file of the form tmp{solution ID}.txt. This is to prevent
solution from acquiring the file handle of a still-being-written
fitness file, because our multiprocessing is ad-hoc and lock free.
Then, the file is moved to fitness{ID}{world ID}.txt.
    When converting to a multi-world simulation, it did not occur to
me to add the world ID to the tmp file. This was a mistake, and
created a data race where A & B variants of a solution (running in
parallel!) would attempt to write/move the same file. Thus a data race
was born.
    Now, we write fitness to tmp{solID}{world ID}.txt, and then to the
appropriate fitness file