Calling this O-K for Deliverable 3
	Currently, the robot is not performing exactly how I want. In
most runs it plots a weird skittering path along the x axis in both
worlds but does react more strongly to bumping into the block, & does
move away from the block when it hits it. The project itself is
somewhat ambigious on exactly how strong the behavior needs to be
	 I also met with Krishna (4/25) and he seemed to think I was
OK to move on to A/B testing. He offered some suggestions for
improving the behavior but jumped pretty immediately to suggesting A/B
testing variants, I did repeatedly press the "am I OK to move on to
A/B testing" issue & he was like if you're getting behavior you
want *on average* then yea, and like on average it's pretty close to
the behavior I'm seeking so I'm going to go ahead.
    	 I may revisit tweaking this in the future, but I only have so much
time to collect data & so on.